+++ src/backend/app/projects/project_logic.py
@@ -1,3 +1,4 @@
+import os
 import json
 import uuid
 from loguru import logger as log
@@ -6,10 +7,13 @@ from fmtm_splitter.splitter import split_by_square
 from fastapi.concurrency import run_in_threadpool
 from psycopg import Connection
 from app.utils import merge_multipolygon
+from app.projects.image_processing import TaskProcessor
 import shapely.wkb as wkblib
 from shapely.geometry import shape
 from io import BytesIO
-from app.s3 import add_obj_to_bucket
+from app.s3 import add_obj_to_bucket, list_objects_from_bucket, get_file_from_bucket
+from concurrent.futures import ThreadPoolExecutor
+
 from app.config import settings
 
 
@@ -133,3 +137,48 @@ async def preview_split_by_square(boundary: str, meters: int):
             meters=meters,
         )
     )
+
+
+def process_imagery_in_nodeodm(
+    images_folder: str, project_name: str, odm_task_id: uuid.UUID
+):
+    token = ""
+    task_name = project_name
+    images_dir = images_folder
+    node_odm_docker_url = settings.NODE_ODM_URL
+    task_processor = TaskProcessor(
+        token, task_name, images_dir, node_odm_docker_url, odm_task_id
+    )
+    task_processor.process()
+
+
+def download_object(bucket_name: str, obj, images_folder: str):
+    if obj.object_name.endswith((".jpg", ".jpeg", ".JPG", ".png", ".PNG")):
+        log.info(f"Downloading image from s3 {obj.object_name}")
+        local_path = f"{images_folder}/{os.path.basename(obj.object_name)}"
+        os.makedirs(os.path.dirname(local_path), exist_ok=True)
+        get_file_from_bucket(bucket_name, obj.object_name, local_path)
+
+
+def process_images_concurrently(
+    project_id: uuid.UUID, task_id: uuid.UUID, odm_task_id: uuid.UUID
+):
+    objects = list_objects_from_bucket(
+        settings.S3_BUCKET_NAME, f"publicuploads/{project_id}"
+    )
+
+    # Construct the images folder path
+    images_folder = (
+        f"/tmp/{project_id}/{task_id}"  # NOTE: local path is set to /tmp/ folder.
+    )
+    project_name = "Testing Project"  # FIXME: Use real Project name here
+
+    # Process images concurrently
+    with ThreadPoolExecutor() as executor:
+        executor.map(
+            lambda obj: download_object(settings.S3_BUCKET_NAME, obj, images_folder),
+            objects,
+        )
+
+    # Call process_imagery_in_nodeodm after images are downloaded
+    process_imagery_in_nodeodm(images_folder, project_name, odm_task_id)
